mise à jour

+++ src/main/java/org/psychohelp/psychohelp/repository/CategorieRepository.java
@@ -3,5 +3,5 @@ package org.psychohelp.psychohelp.repository;
 import org.psychohelp.psychohelp.entity.CategorieTest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CategorieRepository extends JpaRepository<CategorieTest, Long> {
+public interface CategorieRepository extends JpaRepository<CategorieTest, Integer> {
 }
